refactor to use summaries
